requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
Pillow>=9.0.0
numpy>=1.21.0
matplotlib>=3.5.0
pandas>=1.3.0
scipy>=1.7.0